Improved exit codes
Formerly uacme would return 0 (success) when the certificate
is still valid, and 1 for any error.

With this commit, it returns 1 when the certificate is still valid
and 2 for any error. This can be used in cron jobs to take necessary
actions such as restarting services only when the certificate has
been reissued.