use gnutls_decode_rs_value if available
gnutls_decode_rs_value() makes it possible to avoid linking directly
with libtasn1 but is only available in GnuTLS 3.6.0 and later.

In order to keep compatibility with earlier GnuTLS versions the configure
script only checks for libtasn1 if it can't find gnutls_decode_rs_value.

Also taken the opportunity to add -fpedantic to the compile flags,
and eliminated some compiler warnings.