Compatiility with LibreSSL
limitations as of LibreSSL version 3.4.2:

uacme: --must-staple doesn't work. This is because LibreSSL does not
support NID_tlsfeature and therefore it can't add the necessary
extension to the certificate request.

ualpn: doesn't work at all and probably never will. This is because
the TLS implementation of LibreSSL is completely different.