nsupdate.sh: support settings from environment
Currently, nsupdate.sh is usually provided by distributions, but the
only way to configure it is by actually editing it. It would be great to
decouple this file from the configuration (or at least support doing so)
by taking certain settings from the environment.

This patch does so by taking the most common settings from the
environment if set, otherwise using the same values as before. All
variables are taken from UACME_<VAR>, except for DIG and NSUPDATE, where
I felt it prudent to use UACME_DIG_CMD and UACME_NSUPDATE_CMD instead
for clarity.

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>